requirements.txt
appdirs==1.4.4
asgiref==3.2.10
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.9.1
blog==1.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==3.0.8
djangorestframework==3.11.0
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
Markdown==3.2.2
MarkupSafe==1.1.1
pip-tools==5.2.1
pipenv==2020.6.2
psycopg2==2.8.5
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.6.1
pymdown-extensions==7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
Serializer==0.2.1
serializers==0.2.4
simplejson==3.17.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
views==0.3
virtualenv==20.0.27
virtualenv-clone==0.5.4
Werkzeug==1.0.1
